move unused FE modules from shared
