+++ backend/src/game-lobby-service/game-lobby/GameLobby.test.ts
@@ -1,7 +1,7 @@
 // Internal
 import { GameLobby } from './GameLobby';
 import { PokerTable } from '../../game/PokerTable';
-import { createPokerTable } from '@tests/helpers/createPokerTable';
+import { createPokerTable } from '@Tests/helpers/createPokerTable';
 
 describe('GameLobby', () => {
   let gameLobby: GameLobby;
