+++ backend/src/handlers/actions/PokerTableJoinHandler.ts
@@ -4,7 +4,7 @@ import type { PokerTableJoinPayload, PokerTableJoinOutput } from '../../shared/a
 
 // Internal
 import { BaseHandler } from '../BaseHandler';
-import { Result } from '@shared/Result';
+import { Result } from '@Infra/Result';
 import { Rooms } from '../../sockets/Rooms';
 import { GameLobbyService } from '../../game-lobby-service';
 import { PokerTableDoesNotExistError } from '../../shared/errors/PokerTableJoinErrors';
