+++ backend/src/handlers/actions/PokerTableLeaveHandler.ts
@@ -6,8 +6,8 @@ import type { PokerTableLeavePayload, PokerTableLeaveOutput } from '../../shared
 import { BaseHandler } from '../BaseHandler';
 import { Rooms } from '../../sockets/Rooms';
 import { GameLobbyService } from '../../game-lobby-service';
-import { Result } from '@shared/Result';
-import { PokerTableDoesNotExistError } from '@shared/errors/PokerTableLeaveErrors';
+import { Result } from '@Infra/Result';
+import { PokerTableDoesNotExistError } from '@Shared/errors/PokerTableLeaveErrors';
 import { pokerTableLeaveSchema } from '../../shared/api/types/PokerTableLeave';
 
 /**
