+++ backend/src/shared/errors/DB/DBDeleteErrors.ts
@@ -1,4 +1,4 @@
-import { BaseError } from '@shared/Result';
+import { BaseError } from '@Infra/Result';
 
 export class DBDeleteError extends BaseError {
   constructor(public tableName: string) {
