+++ backend/src/shared/errors/DB/DBSelectErrors.ts
@@ -1,4 +1,4 @@
-import { BaseError } from '@shared/Result';
+import { BaseError } from '@Infra/Result';
 
 export class DBSelectError extends BaseError {
   constructor(public tableName: string) {
