+++ backend/src/shared/errors/PokerTableErrors.ts
@@ -1,4 +1,4 @@
-import { BaseError } from '@shared/Result';
+import { BaseError } from '@Infra/Result';
 
 export class PlayerAlreadySeatedError extends BaseError {
   constructor() {
