+++ backend/src/shared/errors/ValidationErrors.ts
@@ -1,4 +1,4 @@
-import { BaseError } from '@shared/Result';
+import { BaseError } from '@Infra/Result';
 
 export class InvalidRequestPayloadError extends BaseError {
   constructor(details: any) {
