+++ backend/src/sockets/Sockets.test.ts
@@ -3,7 +3,7 @@ import { io, Socket } from 'socket.io-client';
 
 // Internal
 import { SocketServer } from '../index';
-import { shutDownServer } from '@tests/helpers/shutDownServer';
+import { shutDownServer } from '@Tests/helpers/shutDownServer';
 import { Sockets } from './Sockets';
 
 let socket: Socket;
