+++ backend/src/tests/helpers/createPokerTable.test.ts
@@ -1,5 +1,5 @@
 // Internal
-import { createPokerTable } from '@tests/helpers/createPokerTable';
+import { createPokerTable } from '@Tests/helpers/createPokerTable';
 
 describe('createPokerTable', () => {
   it('should create a poker table', () => {
