+++ backend/src/tests/mocks/dbMocks.ts
@@ -2,10 +2,10 @@
 import { RowDataPacket } from 'mysql2/promise';
 
 // Internal
-import { Result, ResultSuccess, ResultError } from '@shared/Result';
+import { Result, ResultSuccess, ResultError } from '@Infra/Result';
 import { MySqLInstance } from '../../db/MySql';
-import { DBInsertError, DBInsertDuplicateError } from '@shared/errors/DB/DBInsertErrors';
-import { DBSelectError } from '@shared/errors/DB/DBSelectErrors';
+import { DBInsertError, DBInsertDuplicateError } from '@Shared/errors/DB/DBInsertErrors';
+import { DBSelectError } from '@Shared/errors/DB/DBSelectErrors';
 
 export const mockMySqlInsertError = (table: string) => {
   jest.spyOn(MySqLInstance, 'insert').mockImplementation(async () => await Result.error(new DBInsertError(table)));
